refactor(teacher-calendar): add major improvements
